3874611 : Non ID Based Matching
https://bugs.eclipse.org/bugs/show_bug.cgi?id=387461

Introduces another implementation of EObject matcher which :
- browse the scope, indexing the model elements.
- browse the indexes, finding matches along the way

The indexes are able to return the closest element of an other side.
This is implemented by computing the number of differences we have
between two elements. This is not free regarding CPU cycles but we are
sure to get the optimal result.

The new matcher is plugged in the current implementation by introducing
the ability for the ID based matcher to delegate to another matcher when
an ID is not found.

I named it "ProximityEObjectMatcher" as its based on a distance function
used to compare two EObjects. The current implementation of the distance
function we are using is roughly an "EditionDistance" : it counts the
number of changes you have to make to transform one object to the other.

How do we do that ? By calling EMF Compare's diffing during the match by
the way ! (ok, it may sound twisted). It allows us to fine tune a lot of
thing like ignoring some EStructuralFeatures for the distance, or having
different weight for different features.

No UI is provided to change this so far, its "framework only" but the
default behavior will use IDs when available and fall back to the
Proximity matching on the remaining objects.

Performance-wise the ProximityMatcher is about 10 times slowest than the
ID based one on a decent UML model which have very few differences. Have
a look on the corresponding test case if you want.

By the way all this implementation is fairly provisional.

ps : while coding I introduced a few builder interfaces to configure some
instances.