Change isThreeWay to an attribute
isThreeWay used to be a method browsing the comparison model to see if
some origin has been defined in some ResourceMatch. This could only work
for comparisons having resources and not plain EObjects comparison.

Keeping the same approach would have meant to browse through the whole
comparison model to see if some Match had a reference to an origin. That
looks like quite a bad idea.

So, this commit changes the approach, now the match engine itself set an
ttribute to specify whether its three way or not. Checking for it is now
only an attribute access.