UI Tests with SWTBots
Adds two classes of tests, Ecore and UML.

UML is failing because of an bug in the current trunk.
Ecore is supposed to sucess but failing on my station with a
noclassdeffounderror which makes me think its trying to do something
with EMF Compare 1.x which is installed in my dev environnement.