389810 : Dynamic instance match fails
Content based match is now indexing the instances using a key for each
EClass (string based : EPackage NSURI + EClass name) instead of EClass
java instances.

Also updates parts of the UI which are not using safeEGet( yet and leading
to EMF errors when accessing model elements with a feature from the
other EClass.