Dynamic Threshold for Matching objects by Content
Change the threshold mechanism (maxDistance) so that its no more an
absolute value but it's dynamically computed from a given EObject,
checking the values which are set and using those + some black magic to
set a comprehensive threshold.

The function to compute this max distance is fairly magic but gave
pretty good results so far. I'll now gather more test data to assess the
quality of the match so that I can update this function to a saner
implementation later on while making sure I have the same results.