Bug 390664 - ignore XMI id option does not fully ignore ids
https://bugs.eclipse.org/bugs/show_bug.cgi?id=390664

With this commit the structure based matching is no more comparing URIs
to evaluate the location of an EObject. The problem was some URI could
contain reference to an ID and so even if you want to ignore IDs it
would have an effect on the computation.

It's just a first step though as the EqualityHelper which is indirectly
used by the EditionDistance (through the DiffEngine) is still using
URIs to compare EObjects.