391657: Renaming detected instead of addition / deletion
https://bugs.eclipse.org/bugs/show_bug.cgi?id=391657

Several important fixes :

EReferences were being considered when computing the max distance yet
not during the diff computation. It's no more the case.

Computation of the maxDistance was relying on a single EObject and
depending on which object was passed first could give different result.
It now considers the pair of objects and keep the max maxDistance
between those. As a matter of fact the responsability of the max
distance is now kept into the DistanceFunction.