Potential Merge error with move differences.
If we create a new container, then a new object within this new
container, we will have two differences detected. If we try and merge
the "move" towards the side where it should actually "reset" the move,
the merge action previously was not doing anything (it was "finding" the
wrong target container).

This commit also adds a new "checkMergeState" method that consists of a
long assertion list that checks the result of a merge operation. This
currently uses a static final boolean to enable/disable, but needs to be
changed to a more flexible approach.