add AccessorRegistry extension point and appropriate extensions for emfcomapre core and uml2
Previously, the Accessors (input of panels from ContentMergeViewer) were
created from ICompareInput (ie item from the StructureMerge tree). Now,
ICompareInput delegates to an AccessorFactory. The proper
AccessorFactory is retrieved from a registry. The registry is populated
from Eclipse extensions (see EMFCompareIDEUIPlugin). A factory has a
boolean isFactoryFor(Object) method that returns true if it can handle
the given ICompareInput's target (as it is an EAdapter). It also has a
integer ranking to create a partial order between AccessorFactory.

AbstractEDiffNode subclasses will use the AccessorFactory with the
highest ranking. Now, if extensions are subclassing the core
EMFCompare metamodel and want to provide specific input to specific
viewer, they just have to provide AccessorFactory extensions with
a proper implementation of boolean isFactoryFor(Object) method and to
give the proper ranking. FYI, the rank of core AccessorFactory is 10,
and the rank for UML is 20.