Encapsulate all storage manipulation within the sync model.
The synchonization model no longer exposes the
get(Left|right|Origin)ResourceSet methods as API, nor does it expose
minimize. All manipulation of the storages are now hidden from view, and
the only available API is to createMinimizedScope that directly creates
an IComparisonScope that can be passed as is to EMF Compare.

This was done in order not to expose too much of our innards : the
IComparisonScope should understand that only a limited set of the
resources contained by the left, right and origin resourcesets are part
of the files that need be compared, which is only known to the
synchronization model. This is mainly true for UML : the resource sets
may contain profiles, though they usually don't need to be compared.