Allow content viewers for stereotype application changes.
For stereotype changes, we will display the list of all applied
stereotypes on the current element. However, this is not retrieved
through a standard feature, it can only be done through an API call to
Element.getAppliedStereotypes. This makes the usual computation of the
insertion index fail.

DiffUtil.findInsertionIndex is now visible from the outside so that
we can make use of it in our specialized version of the accessor. We
still have it tagged as @noreference.