We cannot register the changeRecorder ourselves.
This was provoking weird changes descriptions, 'consolidateChanges'
removing some of the changes (for example, if we 'unset' a reference
from Class A to another object, then remove A, the 'unset' change will
be lost, leaving only the object removal... and leading to missing
notifications when undoing.
This was noticed on stereotype application changes.