add load on demand policy extension for UML2.
Static UML profiles must load profiles' resources when manipulating
Stereotypes and StereoTypeApplications. Dynamic profiles do not have
this problem because EPackage defining profiles are contained within an
EAnnotation on the Profile, and then shares the same resource. Static
have to keep a map between the nsURI of the EPackage defining profile
and the uri of the resource containing the profile.