Reduce the number of "refines" differences for diagram extensions.
There is a lot of "noise" differences in diagram comparisons. For each
extension, we were taking all differences "under" a given diff (node
change, edge change...) as a "refinedBy" link. However, we might have
other extensions underneath one another : adding the Shape for a Class
might hide additions of "Property" Shapes. These properties are not
refined by the NodeChange for the Class, but by the NodeChange for each
property.

The new algorithm stops on each extension it detects without adding
sub-differences under those.