Properly detect requirements for the edge changes.
We need to look at the requires/requiredBy links of both source and
target of the changed edge in case we might have some that are
required/requiredBy the extension.

This also modifies the way we're looking for refined diffs for the MOVE
edges... This will fail and must be looked at in more details,
considering both MOVE cases : AttributeChanges (changed a bendpoint) and
ReferenceChanges (changed an anchor).