Detach the DiffCrossReferencer when we remove diffs.
Extensions might need to alter the diff model by adding new Diffs, but
also by removing existing ones detected by the default diff engine. In
such cases, we must remove the diff cross referencer from those
"removed" Diffs in order not to keep references to them.

This will be the case for the UML extension that needs to remove
differences that were detected "duplicated" between a subset and a
superset.