Accurate detection of UML sub-/super-set duplicated differences.
We were previously "simply" ignoring any and all differences detected on
references that were supersets of others. This was not accurate and was
ignoring way too many differences.

We need to ignore diffs on a superset only when the diff is actually
duplicated on one of the subsets. This is now done through a reflexive
access on the supersets' list implementation.