Resource attachment completion.
We now detect resource attachment changes not only for root matches, but
also for cross references. This allows us to detect a wider range of
potential changes, notably the control/uncontrol case.

The implementation is functional for control/uncontrol changes, as well
as addition/deletion of roots in a resource.