394668: Version 2.1 results 'compare' versus 1.3
https://bugs.eclipse.org/bugs/show_bug.cgi?id=394668

Fix match by content regression introduced by the recent changes in the
URIDistance computation.

We were using a "simili uri" in the form of /someRef/indexInRef/otherRef
. Problems are rising when some elements are added/removed in a
collection the indexInRef is shifted and if you're 'unlucky' you'll ends
up matching something (let's say an EAttribute) with another EAttribute
which happens to be at the previous index.

To lower this effect the weight of the location is kept minimum and the
URIDistance is more fine grained (it used to be 0 : same URI, 10 :
different URI)

We are now leveraging the classical fragment mechanism of EMF (available
in InternalEObject) to retrieve the fragment while supporting eKeys
which matters a lot in those case.

The URI Distance is also now more fine grained as it is counting the
number of "hops" required to go from one location to another, leading to
values which will go from 0 to 10 and not 0 or 10 exclusively.