[394668] MatchByContent v1 versus v2
this is a major overhaul of the match by content in compare V2
1- use the parent match to compute the distance
2- use doubles instead of ints
3- fix algorithm un-necessary complexity in the double check

API BREAK : IEObjectMather api had to be changed requiring changes in
every client code defining a specific matcher.
The change is trivial : you used to have to return the list of matches,
now we're giving you a Comparison object and it's your responsability to
add the matches you find in it.