[398402] Regression dating back to c25b90c from 06 december 2012.
These two unit tests were failing for a good reason : we have included
one more difference in the equivalences of this diff because we now
remove duplicate diffs (handling of UML subset/superset... see
UMLDiffExtensionPostProcessor#copyRequirements()).

The tests are now fixed, though the merge was now failing : we have a
difference that is considered "equivalent" to ... one of its
requirements. The merge is fixed through
ReferenceChange#handleEquivalence(), but it will have to be updated when
we tackle bug 398402.