remove the dependency from o.e.e.compare.edit to JFace
I removed the dependency from o.e.e.compare.edit to JFace because this
is not acceptable to depends to UI code from non-UI code. I implemented
my own StyleString structure in .edit and create a converter in .ide.ui
plugin. Color and Font are encoded through EMF URI as in specified in
IItempFontProvider and IItemColorProvider (EMF).

Change-Id: Ibdabfff77d0b3533f3d5fcb50ac1eba9ddb15ae4