[398358] Extension point for filters and groups
Add an extension point to EMF Compare so that users and extenders can 
provide their own filters and groups.

Rewrite the current filters as extension to the extension as examples of
how to contribute these items.

Add new group in case of 3-way comparison with 3 categories : left
side, right side, and conflicts.

Change-Id: I19975b7a0f8f9b66078a6082360716f1d9c0907a