[399361] Performance of the LCS.
Computing the LCS was extremely costly with large input lists. The
implementation has been tweaked in order to reduce the input sets : when
we can find common prefixes and suffixes using one single comparison on
the objects, it is _much_ better to do so than to leave those objects as
a row/column in the LCS matrix.

In the worst case (there's been one diff at the very beginning and one
at the very end of the input sequences), this will only add two
additional comparisons. In the best case, this will remove the whole
matrix computation, and all comparisons it would have entailed.