Simplify logic for overlay and kind of differences display.
The "kind" is used by Team in order to determine whether the "show
ancestor" action is enabled or not. As we were not correctly computing
it, the action was not always enabled for three way comparisons (for
the record, it is enabled if "kind" contains either Differencer.LEFT or
Differencer.RIGHT).

The overlay was using an "always true" boolean, which resulted in us
having 6 unused icons, and an uselessly complex nesting of "ifs". This
has been simplified. Moreover, we no longer use the "incoming conflict"
(red, right facing arrow) icon for the Match elements displayed in the
UI. Conflicts now use a red "in and out" icon such as what's used by
Team compare.

Change-Id: I4ad6524be264969656dd088116b01ca6fbbb0f8d