Bug fixes:
- getFigure() may return null if the edit part does not exist.
- Regression after rebasing: the graphical extensions are not put away
in the same way anymore ! So, the checking of the add of object in a
container where its location changed does not work anymore because the
change location difference is not in the same match anymore. Waiting for
understanding about the first regression, I extended the research of the
match to all the ancestors.