Integrate EMF Compare with Team's model providers.
This commit is the squashed result of merging the branch
logical_experimentation into master. This was done as a new commit
entirely to allow for review. The feature branch may be removed in the
future.

When Team is asked to compare a given file with one of its remote
revision, the intended process is to resolve the whole logical model
that this file is a part of. For EMF models, the "logical model" is the
set of all resources that are referenced from (or that reference) the
compared file. For example, a genmodel file cannot be compared, let
alone loaded, without its underlying ecore file.

This implementation makes use of non-API features of Team (through
reflection and internal accesses) as it needs some locked out features
to work properly with EGit and CVS without explicitly depending on
these two specific implementations of repository providers.

Furthermore, this commit adds specific IDE UI tests in order to make
sure this integration works as designed. It thus re-enables these tests
for the build, removes the SWTBot dependency along with the two
non-functional SWTBot tests that previously lived in the ide.ui.tests
plugin. This adds build-time dependencies towards EGit and JGit.

Change-Id: I9b1c47d79516037df1cd2495abb061586ca6a4cf