Improvement of computation of the decorators after merging or undo/redo.
All the decorators are removed after a merge of a difference or after
undo/redo.
This allows to compute again the decorators because the merge of an
object can involve a shift of objects (e.g. merge of the ADD of element
in a list owning other differences or merge of a coordinates change of a
container owning other differences too...)