Bypass logical model when no subscriber is available
Complete logical model tests so that we can exert a number of possible
paths for the resolution through unit testing... and try for these
tests to be accurate (we cannot make sure we've been fed the proper
revisions, but we can make sure the number and sides of detected
differences match what we expect).

Fix all resolution failures highlighted by these new tests.

Try and handle the "exception case": repository providers are supposed
to query the model providers extension point to make sure that a file
can (or cannot) be compared alone. When they do not, we are only fed
with the two (or three) initial revisions to compare, and we have no
way to load the cross-resource references of these initial versions. We
now make sure not to even try and follow these references by only
loading what we were fed and creating a minimized scope on that alone.

Change-Id: Id942b75206bf2449249061efa5f61bea8f062d2c