[405000] Performance enhancements on the LCS and its use
Consider that we need the LCS between two lists L1 and L2 of size M
and N respectively.

Memory-wise: we now use a short-typed score matrix instead of the
previous int types. The maximum score will be Max(M, N), which in most
cases will never get greater than Short.MAX_VALUE. A special case has
been inserted to fall back to an int-typed matrix in case this
threshold is crossed... but that would mean we need to instantiate a
score matrix which size is _at least_ int[2^15][2^15]. The cost in
memory of such a matrix is probably too high for any desktop computer.
This resulted in two copy/pasted implementations of the LCS; we need to
extract these in there own implementation of an extracted LCS
interface, but that will only make real sense when we tackle a more
practical implementation of the LCS.

Time-wise: the DefaultDiffEngine no longer iterates over and over on
the LCS of the compared lists. Since we know the LCS elements appear in
the same order as in said lists, we can safely iterate once at a time
and use a single comparison instead of the previous full iteration. In
terms of comparison count, we are now down to a maximum of (M+N)
instead of the previous constant (M+N)*(LCS size).

Bug: 405000
Change-Id: I116598ea0061e9df4d694b4f3eb9dda9320fac26