Avoiding Complexity Explosion
This commit introduces a new mechanism to limit the CPU complexity of
the
match by content algorithm when there are many EObjects of the same type
to match.

It introduces a new interface EObjectIndex's might implement
: MatchAheadOfTime. If an index implements this interface it
will be asked by the Matcher, from time to time, to provide elements
to be matched whereas the whole scope has not been processed yet.

The default implementation in ProximityIndex checks for the number of
elements kept in its storage and depending on this number decides to
return elements to match ahead or not.

This effectively cut the CPU complexity on models having for instance
10K elements of the same type, this has a drawback though : when this
mechanism is triggered, we can no longer guarantee the match will be
the "best match ever" as some elements have been matched before all the
candidates where known.

ProximityEObjectMatcher.NB_ELEMENTS_BETWEEN_MATCH_AHEAD = 10 000
means every 10 000 elements the matcher will ask the indexes for
elements to match ahead.

ProximityIndex.SEARCH_WINDOW = 1000 means a given index (which holds
all the elements of a given type) will only return elements to match
ahead if it has more than 1000 elements to match already.

These values means the match ahead will have no impact on models which
have less than 10 000 elements or which don't have more than 1000
elements of the same kind.

Matching guava.uml (50K elements) goes from 50s to 11s to match the model 10 times.

Change-Id: I0f9c6f536336aea1742389c991e3c672a67a154c