Replace Guava's loading caches by HashMap
We've seen and measured Guava's cache is costly. It records usage
statistics (and we can't disable that in the version we are using)
andthe implementation itself does not scale well (yet) with big
maps:
LoadingCaches behave as LinkedList, displaying very poor performances
until Guava 12.0.1, more details on
https://code.google.com/p/guava-libraries/issues/detail?id=1055.

This commit replaces Guava's by a simple HashMap with a
custom getOrCompute() method.

Change-Id: I4bfc8a3ccc56d361b66b7aa362ff125f0f0f6751