Filling of the implication link with UML subsets.
Like equivalences, the merge of a difference may imply the merge of
other ones. But, contrary to equivalences, this implication link is
unidirectional. So, if A implies B, B does not imply A.
In the context of UML, references can be defined as subset of other
ones. So, the merge of the add of a reference which is subset of
another one implies the add of this last one. In the same way, the
delete of a reference which is superset of other ones implies the
delete of these last ones. Note that some superset references can be
defined as 'union' (that is to say that each superset reference is in
one of the subsets). Union supersets are derived, so they are not
managed by EMF Compare.
The implications are set in Diff#implies and Diff#impliedBy.
The reference merger of EMF Compare scans these links to mark as merged
the impacted differences without merging them (because the merge
itself will be managed by the 'meta-model').

Change-Id: If3a62076c3569b1ba0aa4628c657765cd8ad7711