Fix requirement link between a unit difference and a macroscopic change.
It is required to build a requirement link between a unit difference
and a macroscopic change to delete all impacted objects 
(in a macro object) by a delete of a unit object.

For example, the delete of a UML Class involves the delete of the 
"element" reference coming from a GMF Node. This last difference is 
embedded in a macroscopic change (the delete of the set of the 
related GMF nodes representing a graphical object). 
So, the delete of the unit UML Class requires the delete of the 
macroscopic graphical object.

This link is built in the same way as the one between 2 macroscopic 
changes.

Change-Id: I2dac280c573ac7805a0080b0044c749bfab5fc02