Extract and extend comparison scope builder, minimizer, and resolver
This extracts the comparison scope building into new APIs, preparing
the code for extensibility. IModelResolver and IModelMinimizer have
been introduced to allow clients to customize how the logical model is
resolved, and how it is to be reduced to the minimal set of change
candidates before comparison.

This also introduces a new resolving scheme that looks up a whole
IProject for references to a given IFile instead of searching only for
said IFile's "downward" dependencies. This allows us to see the "root"
of the logical model even when starting from a "leaf", if said root is
in the leaf's containing project.

Change-Id: I02c90eed348ffec620719c032520a913d144f86f