Prevent logical model from containing multiple starting points
During local comparisons, the logical model is computed for all three
selected resources. If there is one resource in the workspace that
reference (or "is referenced") by more than one of the compared
resources, we'll end up with a logical model that will be identical
for all of these resources.

We now constrain the graph traversal to avoid iterating over other
"compared resource" (say "R2") when we already started from one (R1).
This keeps us out of R2's own dependency graph. Afterwards, we remove
the intersection of the graphs before creating the actual comparison
scope.

Change-Id: I97438e644fdea423f3af08be1e09784c01423e91