A macroscopic change for differences on UML DirectedRelationship.
Like changes on UML associations, a macroscopic change must be created
for changes on DirectedRelationship elements.
Without this macroscopic change, the add of an incoming reference to 
the DirectionRelationship (Generalization for example) will add the 
DirectionRelationship itself but not the potential outgoing references 
from it (Generalization#general for example) which are integral part of
the related concept. So, the add of a graphical edge which represents a
UML generalization will involve an inconsistency in the business UML
model. That's why it is required to encapsulate all the differences 
related to a DirectionRelationship change in a macroscopic one. The 
requirement link will be automatically computed between this one and 
other ones.  

Bug: 406405
Change-Id: I6c624a913bcf993b666421ba8fae868e0b2ae753