Detection of change of profile version.
For UML dynamic models, we checks we have the same profile version
between the 2 sides. For that, we compares the URI (fragment) of the
generated EPackage (in the profile) referenced from each profile
application. If it is different, then the profile application bases on
a different profile definition and a RuntimeException is thrown.

To test this, a refactoring of the UML test framework was needed to
correctly initialize JUnit tests with the registration of the EPackage
(profile) and the map to find the profile location from this EPackage
in static cases and not to do that in dynamic cases.

The end-user is warned from the profile version difference by a message
dialog.

Change-Id: I67e7ef8b90aca100440818bcf76ef87cfd036be9