[416115] The edge phantom is not attached to the node phantom
In case of a delete/add of an edge while the connected nodes have 
respectively moved and been deleted, the computed phantom related to 
the edge is drawn from the moved node (OK) but not to the phantom 
related to the deleted node.

Indeed, no edit part exists, in the "phantom" side (the phantom 
related to the deleted node is just a GEF figure without edit part).

A GEF edit part has to be created. It builds an invisible figure with 
the same bounds as the phantom and it is attached to the edge 
phantom edit part (as target) in order that the edge be connect to the 
node phantom.

Bug: 416115
Change-Id: I9ff902b4a131f689ef671574e4de6b9792853ab2