Add overlay support for any object.
It adds a transparent image as overlay to any other image. It is
important when the first display element in the SMV is a group, it has
to have the proper size for diff and match to be displayed later.
Without this decorator, the image in the tree will all have the size of
an image without overlay and then the diff with markers's image will be
squized.

Change-Id: I756f99f0344b127a12fd37e173ef6bff665fd494