Improve robustness against error
Whenever an error is raised during the model resolving or the comparison
computation, diagnostics are created and returned to the launcher job so
that it can display the diagnostic to the user. It reuses parts of EMF
to display the diagnostics à la Error Log. It will be the basis for the
fallback to standard text comparison.

Change-Id: I0c81a819816b5c8366035333ea5fe5fcd77d93e2