Fix random occurences of resolving issues
This fixes an issue with the local model resolving which
ignored the "in sync" files even though that is the
responsibility of the IModelMinimizer.

One of the resolution end signals was coming too late,
which meant that we could end up querying the dependencies
of a resource before the resolution ended.

Finally, the remote traversals are also constructed through
multiple threads, but the StorageURIConverter wasn't
properly guarded against concurrent accesses, making for
random but very frequent resolution failures for the remote
sides of a comparison.

Change-Id: I8d70072974f0b948a22c072054dd8710f65c1f27