The logical model should find incoming cross-references
This makes sure that the logical model of an EMF resource
is capable of discovering incoming cross-references
(cross-references added in a remote version or deleted from
the local one). Basically, if I have two files (F1 and F2)
that are not linked together locally, but that are (or
were) part of the same logical model in either the remote
or the original versions, then they should be considered
together locally as well, lest we detect a lot of false
differences (everything deleted, for exemple).

Change-Id: I3c3c62b617b27017c37a2b233f8b9980749bb749