UI improvement: Prevents freezing UI during groups building.
Refactors the structure merge viewer in order prevents freezing during
groups building. This has been done by creating a external job for each
new group provider construction. While the groups is building up a
pending notification is displayed to the user. When the group is built
the pending notification is removed from the tree and replaced by the
correct children. Once a group has been built for a comparison no more
pending element will be diplayed for this group.


Change-Id: I5d1fe459bc26195e48f4af0011407cae604e32b3
Also-by: Mikaël Barbero <mikael.barbero@obeo.fr>
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>