[413520] Fixes IndexOutOfBoundsException on merging opposite references
The IndexOutOfBoundsExceptions occurred when a one-to-many reference has
been changed, whereas the original object on the single-valued side of
the one-to-many reference has been deleted. During the check for the
ordering of the target values affected by equivalent changes (in this
case, changes of opposite references) in
ReferenceChangeMerger.internalCheckOrdering(ReferenceChange, boolean),
the source and target container and, in the next step, their values at
the feature constituting the multi-valued side of a one-to-many
reference are obtained and their LCS is computed to finally move the
respective value. However, when the original object on the source side
has been deleted, the source container is null, and the values at the
respective feature are empty. DiffUtil.findInsertionIndex will still
return 0, so that EList.move(index, value) is called, whereas value is
not available in the list, which ultimately causes the exception.

This commit prevents the call of
ReferenceChangeMerger.internalCheckOrdering(ReferenceChange, boolean)
for any difference that is not an addition to a multi-valued feature in
the current merge (thus, not a move or a delete), since only for such
differences a check of the resulting order should be necessary.

Signed-off-by: Philip Langer <planger@eclipsesource.com>
Bug: 413520
Change-Id: I363b5c12538f847cf0d0a359000969a2e09f54f4