Fix merge-related UI feedback of the Structure viewer
This introduces API for the mergers to describe what they
consider to be implications of a given merge operation,
whether it be diffs that need to be merged along the
selected one or diffs that need to be rejected because of
the seleced one.

IMerger2 allows clients to describe their own algorithm,
AbstractMerger implementing the default navigation.

Change-Id: I4ac1f0590b19ad81934c0244e1d962feae88a2c9
Signed-off-by: Laurent Goubet <laurent.goubet@obeo.fr>