[441258] Prevents IllegalArgumentException on merging moves
When the target reference of a move is multi-valued, the
ReferenceChangeMerger invokes DiffUtil.findInsertionIndex(Comparison,
Diff, boolean), which in turn throws an IllegalArgumentException, if the
feature to which the ReferenceChange refers is single-valued. This,
however, can happen when moving an element form a single-valued
reference to a multi-valued reference and doing a right to left merge.

To prevent this IllegalArgumentException, we make sure that only the
target reference is checked for being multi-valued, irrespectively of
whether the difference refers to the single-valued or multi-valued
containment reference involved in the move. Therefore, we had to
refactor out methods for getting the target feature and container.

Signed-off-by: Philip Langer <planger@eclipsesource.com>
Bug: 441258
Change-Id: I3e7b6609145023a453d890259809d3d332af7f38