[442355] Corrects label on "ReplaceWith" dialog.
The computation of the label needs to be aware of the visible elements.
Previously it was done using the content provider. However now it's not
possible anymore since the content provider is deferred. To compute the
list of visible items the implementation now rely on the swt tree
content that is to say on the TreeItem inside of it. The
updateTreeItemMappings forces the tree to be filled.

Change-Id: I0b0616dc6bac894dcda53d02c2ba971bb69ed34e
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>