Copy all commands didn't work with UML
... and cascading elements. If the only visible difference
in the compare editor is an UML "refining" difference, then
trying to "copy all non-conflicting" will result in a
working merge but an unsaveable model (editor wasn't set as
"dirty") and an empty content pane.

This should probably be investigated more : why are we
relying on the "setSelection" post-merge to set the editor
dirty? This may fail in more cases than this one.

Change-Id: Ic5b8d73d8fb2bbdfbf0c59efdede0b76accfdc37