[443330] Avoid getting a "Profile version change" error.
Previously in order to check that the version of a profile had not
change the program was using URI comparison in order to compare that the
EPackage pointed by the two URI were identical.  In some case, when the
profile was referenced using PATHMAP, it happens that the uri comparison
fail whereas the to uri point to the same object (One of the uri was
referencing the object using a pathmap whereas the other was using the
resolved uri). Normalizing uris before comparison corrects this problem.

Bug: 443330
Change-Id: I046dad5edf68aa9a59dbe347a8302ab91b49357e
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>