[443420] Handle not integrated dynamic profile.
This modification force dynamic integrated profile to be loaded. In the
previous implementation only the profile registered against the the UML
profile extension point were loaded. However it does not handle
correctly the case of "fully" dynamic profile. That is to say profiles
that are directly referenced from a UML model without being registered
in the uml profile registry. Now an approximation will decide whether of
not an URI references a profile using the convention ".profile.uml" as
file extension. Registered profile that does not use this convention
will still be loaded.

Bug: 443420
Change-Id: Icde513b9dd52501953194d8875622231f0de99d1
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>