Consider potential duplicates in String similarity
We need to ensure that "aaa" is never considered to be equal to "aaaa".
The previous unit tests were failing to notice the logic error since we
were only comparing "aa" and "aaaa"... whereas "aa" is less than or
equal to two characters long, falling into a special case within
DiffUtils.

Bug: 444712
Change-Id: I0e81e58eadc288f5a79daac75e245f33ca655119