Modify merge to handle implications properly
This introduces an overhaul of the merging process as a
whole in order to properly handle implications (a diff
implied by another does not need merging in itself, we
only need to merge the implying), most notably for the UML
comparison support.

The same algorithm is now used to merge differences in a
batch (or to simply merge implications, results and
equivalences) and to display the impact analysis of a merge
in the user interface (what will be merged and what will be
rejected when I merge such or such difference).

Change-Id: I39f69c05f94b5f18cb3a3c3b1f7089700f138282