Pre-merge from EMF logical merging
This also introduces an in-depth impact analysis facility
that allows EMF Compare to automatically pre-merge
everything that can be (i.e everything that is not in
conflict or does not depend on a conflict resolution). This
is notably used within the logical merging, but could also
be used for the "merge all non-conflicting" actions of the
user interface.

Change-Id: I236cd86a32b58c4e927ec43753f354a186af7ccb