Prevents loading twice the same model in a NotLoadingResourceSet.
The previsous getResource(IStorage,boolean) method did not check if a
resource was already loaded. It may happen that the same resource was
created twice in the same ResourceSet (with the exact same uri).

This commit also changes the way IStorage are loaded in the resource
set. The IStorages are now loaded using their uris. The resource set is
aware of the resources that should be loaded from a IStorage. 

Change-Id: I5fe2f837c7a0c3ff522ee3d382e4b3dc7fc4b9d9
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>