[446947] Adds support for three-way merging multi-line String attributes
Changes applied to String attributes, which might for instance contain
multiple lines of code in a model (e.g., OpaqueBehaviors in UML), will
now be checked whether concurrent changes in the context of a three-way
diff are mergeable using a three-way line-based merge algorithm. If yes,
no conflict will be raised and they will be merged line by line.

Therefore, this commit introduces a copy of the diff_match_patch class
R106 and a ThreeWayTextDiff utility, which extends the capabilities of
diff_match_patch to support three-way diffing and merging of text (as
opposed to two-way diffing only).

Bug: 446947
CQ: 8824
Change-Id: Ifc7a3144461f7d47750b90a8098d565279ee826d
Signed-off-by: Philip Langer <planger@eclipsesource.com>