[448464] Source list incorrectly determined when undoing changes
When computing the insertion index during a three-way merge of a change
that affects multi-valued features, the source list of a diff was
incorrectly determined in DiffUtil#getSourceList(Comparison comparison,
Diff diff, boolean rightToLeft) if the current merging undoes a change. 
With the current fix, the source list is obtained from the _target
container_ retreived from the _match in the origin_, as opposed to the
list obtained from the _match in the origin_.

The commit also updates the tests that are reaching the now modified
code.

Bug: 448464
Change-Id: I3c53b9ff533b1466d5fddce5f91f5b6de38d9fe8
Signed-off-by: Alexandra Buzila <alexandra.buzila@gmail.com>