[450949] Avoid shortcuts in DefaultDiffEngine for certain FeatureMaps
Changes in FeatureMaps should be merged via FeatureMapChanges in the
FeatureMapChangeMerger. In cases of added / removed elements the
DefaultDiffEngine prevents the calculation of required FeatureMapChanges
since they only affect attributes of those elements. This leads to
ordering problems since the ReferenceChangeMerger can not handle
FeatureMaps properly. This fix removes the calculation prevention for
these cases.

Includes testcases for this issue.

Bug: 450949
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Change-Id: Ie21a54c2e08b27bcc421f1c28d42e542407b990d