The ConflictMerger ranking should be above that of most others
As it was, the conflictmerger was nerver called for custom diff types
such as UML reference changes, which were merge through
UMLReferenceChangeMerger. This lead merge operations through the UI to
merge one side of the conflict while ignoring the other (instead of
rejecting it). The UMLReferenceChangeMerge (or other custom mergers)
will still be called, but only through the proxy of the ConflictMerger.

Change-Id: I0625b3c3b96f967c8b86edee360cc6d1b355c28c